pyproject.toml
[tool.poetry]
name = "model-api-server"
version = "0.1.0"
description = ""
authors = ["centonhuang <centonhuang@tencent.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
poetry = "^1.8.2"
pydantic = "^2.6.4"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
loguru = "^0.7.2"
numpy = "^1.26.4"
tiktoken = "^0.6.0"
cachetools = "^5.3.3"
sse-starlette = "^2.1.2"
flake8 = "^7.1.0"
sentence-transformers = { version = "^2.6.1", optional = true }
vllm = { version = "0.4.1", optional = true }
mlx-lm = { version = "^0.15", optional = true }
autoawq = { version = "^0.2.5", optional = true }
auto_gptq = { version = "^0.7.1", optional = true }

[tool.poetry.extras]
reranker = ["sentence-transformers"]
embedding = ["sentence-transformers"]
vllm = ["vllm"]
mlx = ["mlx-lm"]
awq = ["autoawq"]
gptq = ["auto_gptq"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
